Detangle APDU code from HID implementation
